refactor: move game package in internal
