+++ internal/game/service.go
